# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Finwiz",
        "description": "Finwiz API documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/future-goal": {
            "get": {
                "tags": [
                    "Future Goal"
                ],
                "summary": "Get user's Future Goal with Categories",
                "operationId": "c0a513659b243c17353036da98fc2475",
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Future Goal"
                ],
                "summary": "Create a future goal",
                "operationId": "045db8deaeaa0b9157a78a9e40b04583",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "purchase_name": {
                                        "type": "string",
                                        "example": "Name"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": "111.5"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "example": "High"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-09-16"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/user/future-goal/{id}": {
            "get": {
                "tags": [
                    "Future Goal"
                ],
                "summary": "Get a user's Specific Future Goal By ID",
                "operationId": "8d038676eb12e708ff61294ec5864792",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user future goal to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Future Goal"
                ],
                "summary": "Update a user's Future Goal by ID",
                "operationId": "faec6dd82d371b9df87752e017b07865",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user future goal to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": "11"
                                    },
                                    "purchase_name": {
                                        "type": "string",
                                        "example": "abc"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": "10.5"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "example": "High"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-09-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Future Goal"
                ],
                "summary": "Delete a user Future Goal by ID",
                "operationId": "b5ec0378393e49c364696de899065017",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user Future Goal to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/user/show-auth-credentials": {
            "get": {
                "tags": [
                    "Investment Holding"
                ],
                "summary": "Get user's Auth credentials",
                "operationId": "afea1a7b90cf0470e78ee62f6de284ba",
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/user/show-investment-holding/{accountId}": {
            "get": {
                "tags": [
                    "Investment Holding"
                ],
                "summary": "Get a user's Investment Holding Records",
                "operationId": "dde16c749110f10c8e3c8286644b1a4d",
                "parameters": [
                    {
                        "name": "accountId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/user/notifications": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Get user notifications",
                "operationId": "1bcf48368acee0d64c3ab0b731bc6211",
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/user/read-notifications": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Get user read notifications",
                "operationId": "c30a95331a15aac7b822acdf5ea50f99",
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/user/unseen-notifications": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Get user unread notifications",
                "operationId": "b76ec8a329092606403615af79a272d3",
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/user/mark-read-all-notifications": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Mark all unread notifications as read",
                "operationId": "081d54b90fc4a84b19d4dbc4e48e34e9",
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization ",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearer_token": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization ",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}